feat: consolidate models
